requirements.txt
numpy==1.16.4
requests==2.24.0
beautifulsoup4==4.9.3
Pillow==8.1.0
